tests: mailbox: move all tests to one file
No need to split this into multiple files, just put everything into one
file.
